+++ LinkEngine/GameTypes/Dungeon/Dungeon.cs
@@ -1,0 +1,14 @@
+﻿using LinkEngine.WorldGen;
+
+namespace LinkEngine.Dungeon
+{
+    public class Dungeon
+    {
+        DungeonGeneration generator = new DungeonGeneration();
+        public Dungeon(string _name, int numOfRooms, Biome biome, int playerX, int playerY)
+        {
+            generator.BuildMap(numOfRooms, biome, playerX, playerY);
+        }
+    }
+} 
+