+++ LinkEngine/GameTypes/Dungeon/DungeonItem.cs
@@ -1,0 +1,10 @@
+﻿namespace LinkEngine.Dungeon
+{
+    public class DungeonItem : Entities.Item
+    {
+        public DungeonItem(int _id, string _name, string _namePlural) : base (_id, _name, _namePlural)
+        {
+
+        }
+    }
+}