+++ LinkEngine/GameTypes/RPG/Engine/Gameplay/Equipment.cs
@@ -1,11 +1,9 @@
 ﻿namespace LinkEngine.RPG
 {
-    public class Equipment : Entities.Item
+    public class Equipment : RPGItem
     {
-        public int MinimumDamage { get; set; }
-        public int MaximumDamage { get; set; }
-        public int MaximumDefense { get; set; }
-        public int MinimumDefense { get; set; }
+        public int StrengthBoost { get; set; }
+        public int DefenseBoost { get; set; }
 
         public int Slot { get; set; }
         public bool IsEquipped { get; set; }
@@ -22,30 +20,24 @@
             Equipable = true;
             Slot = equ.Slot;
         }
-
-        
     }
     public class Weapon : Equipment
     {
-        public Weapon(int _id, string _name, string _namePlural, int _minDamage, int _maxDamage, int _cost, int slot) :
+        public Weapon(int _id, string _name, string _namePlural, int strength, int _cost, int slot) :
             base(_id, _name, _namePlural, _cost)
         {
-            MinimumDamage = _minDamage;
-            MaximumDamage = _maxDamage;
+            StrengthBoost = strength;
 
             Slot = slot;
             Equipable = true;
-        }
-
-        
+        }        
     }
     public class Armor : Equipment
     {
-        public Armor(int _id, string _name, string _namePlural, int _maxDefense, int _minDefense, int _cost, int slot) :
+        public Armor(int _id, string _name, string _namePlural, int defense, int _cost, int slot) :
             base(_id, _name, _namePlural, _cost)
         {
-            MaximumDefense = _maxDefense;
-            MinimumDefense = _minDefense;
+            DefenseBoost = defense;
 
             Slot = slot;
             Equipable = true;